requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pdf2image==1.16.3
Pillow==10.1.0
requests==2.31.0 